add new Quesion Chalange

+++ EFCORELEARNING/Models/OrderDetail.cs
@@ -5,7 +5,7 @@
         public int OrderID { get; set; }
         public int ProductID { get; set; }
         public int UnitPrice { get; set; }//money
-        public int Qty { get; set; }//small int 
+        public short Qty { get; set; }//small int 
         public int Discount { get; set; }//numeric
         //many-many
         public Order? OrderOfOrderDetail { get; set; }
